refactor: architecture merge — SavedData buildings, ItemBank warehouse, async BlockInteractOp, ritual visuals
## Architecture Changes
- Building system: replace BlockEntity with BuildingSavedData (Level SavedData)
  - New: BuildingState, BuildingSavedData, BuildCompleteListener, BuildingBreakHandler, BuildingInteractHandler
  - Removed: AbstractWandscapeBE, EarthNodeBE, ForestNodeBE, GrandTowerBE, TownHallBE, WandscapeBuildingBlock, BlockPlaceHandler
- Warehouse system: replace WarehouseBE with ColonyItemBank (SavedData)
  - WarehouseManager.addResource() via ColonyItemBank instead of WarehouseBE
- BuildingApiImpl rewritten to operate on BuildingState instead of BEs

## Node Gathering → BlockInteractOp
- New: WandscapeBlockInteractExecutor (async countdown + thenRun callback)
- BlockInteractOp: configurable channelTicks + manaCost (unlike RitualOp hardcoded)
- node_gathering moved from RitualOp to block_interact action "gather"
- BuildingTaskSource.supplyNodeBuildings() auto-enqueues gather work
- BuildingConfig.NodeConfig with blueprint, element, amountPerHarvest, channelTicks, manaCost

## Async RitualOps
- WandscapeRitualOps: PendingRitual queue + tickAll() countdown for all rituals
- RitualOp.params for functional data (element, amount) — channelTicks hardcoded per RitualId
- Mana consumption fix: consume before execution (both sync and async paths)

## Visual Distinction
- TaskExecutor.currentOpKind: "ritual:self_teleport", "block_interact:gather", "transform"
- WandscapeNpcRenderer: ritual → rotating magic circle (ENCHANT particles); others → wand beam
- Magic circle supports per-ritual styles via ritualId dispatch

## NPC
- WizardHatLayer + WizardHatModel for hat decoration
- 17 wizard skin variants with auto-detection
- WandscapeNpc: DATA_OP_KIND synced field for visual dispatch

## Docs
- Updated docs/05, docs/17 to reflect all architecture changes
- New docs/27-multiblock-refactor-analysis.md

+++ src/main/java/com/wsteam/wandscape/core/component/ManaPool.java
@@ -6,19 +6,19 @@ package com.wsteam.wandscape.core.component;
  */
 public class ManaPool {
 
-    private int current;
+    private float current;
     private final int max;
-    private final int regenPerTick;
+    private final float regenPerTick;
 
-    public ManaPool(int current, int max, int regenPerTick) {
+    public ManaPool(float current, int max, float regenPerTick) {
         this.current = Math.min(current, max);
         this.max = max;
         this.regenPerTick = regenPerTick;
     }
 
-    public int current() { return current; }
+    public float current() { return current; }
     public int max() { return max; }
-    public int regenPerTick() { return regenPerTick; }
+    public float regenPerTick() { return regenPerTick; }
 
     /** Regenerate mana, capping at max. */
     public void regen() {
@@ -26,14 +26,14 @@ public class ManaPool {
     }
 
     /** Try to consume mana. Returns false if insufficient. */
-    public boolean consume(int amount) {
+    public boolean consume(float amount) {
         if (current < amount) return false;
         current -= amount;
         return true;
     }
 
     /** Add mana (e.g., from external source), capping at max. */
-    public void add(int amount) {
+    public void add(float amount) {
         current = Math.min(current + amount, max);
     }
